requirements.txt
Flask==3.0.3
pymongo==4.7.3
bcrypt==4.1.3
python-dotenv==1.0.1
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.7
